backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1
